Started to implement a new design to the application
Starting to change the theme of the application to something more 'Leosac'. Removed the old icon to some brand new icon. Also starting to add an option in which you can modulates the style of the application, but since there is no known method for it, it is suspended. Will also work on the logo url, thnaks to that, ou just have to put your logo in /assets/yourApp, and follow the name given for each logo. You have to set the config file with your path for the logo too.