Started to separate the gpio model from the main application in an ember addon. not an engine, the engines is too big
If I am not using an ember-engines, it is because the ember addon is lighter. It doesn't need route for example. If we are separate this model from the application, it is because the ember-engines wiegand cannot catch the gpio model, because there is no direct link between two model. We have to make and addon to make a link between the app and the engines. Currently struggling with it.