Added a new route to the application: entry-point.
This route will set the Leosac Address when the user whant to use it from leosac.com. The user will set an address in a form on leosac.com, and then when he validate the form, he will be redirected to to this route, and at the end of the url, there will be the address. Then this route will catch the leosac address, and set it so that the application will have a new server address.