Modified the audit-log service.
The unloadAll method doesn't work since the update of ember-data from the version 2.12 to the 2.14.
Instead of the unloadAll, we do a peekAll on each audit type, and then unload each record of every audit.

This is a bit hackish, but since the issue with unloadAll is not resolved, we have to use this instead.