Added a new model to the application: leosac-builtin-access-point
This model will herit from the access-point model. There is no backend to this model, so there is still no adapter. We also can't save a record of this model because of that. The aim of this model is to have a builtin access point, that will allow us to do action when an id is recognized by a reader.